Rework middleware to abstract common response body parsing logic
