+++ src/stac_auth_proxy/middleware/__init__.py
@@ -2,17 +2,16 @@
 
 from .AddProcessTimeHeaderMiddleware import AddProcessTimeHeaderMiddleware
 from .ApplyCql2FilterMiddleware import ApplyCql2FilterMiddleware
+from .AuthenticationExtensionMiddleware import AuthenticationExtensionMiddleware
 from .BuildCql2FilterMiddleware import BuildCql2FilterMiddleware
 from .EnforceAuthMiddleware import EnforceAuthMiddleware
 from .UpdateOpenApiMiddleware import OpenApiMiddleware
 
 __all__ = [
-    x.__name__
-    for x in [
-        OpenApiMiddleware,
-        AddProcessTimeHeaderMiddleware,
-        EnforceAuthMiddleware,
-        BuildCql2FilterMiddleware,
-        ApplyCql2FilterMiddleware,
-    ]
+    "AddProcessTimeHeaderMiddleware",
+    "ApplyCql2FilterMiddleware",
+    "AuthenticationExtensionMiddleware",
+    "BuildCql2FilterMiddleware",
+    "EnforceAuthMiddleware",
+    "OpenApiMiddleware",
 ]
